+++ backend/app/utils/bs4/preprocess.py
@@ -1,13 +1,18 @@
 import os
 import requests
 from bs4 import BeautifulSoup
-from app.utils.bs4.keyword import find_og_info, find_keyword
+from app.utils.bs4.info import find_og_info
+from app.utils.bs4.keyword import find_keyword
 from app.utils.thumbnail.image_resizing import *
-import time 
+import time
+
+from decouple import config
 
 skip_txt = {"", " "}
 skip_word = {"수", "하는", "이", "더", "한", ""}
-file_path = "../data/articles.txt"
+file_path = "app/data/articles.txt"
+
+DEFAULT_IMAGE_URL = config("DEFAULT_IMAGE_URL")
 
 # ToDo : ChatGPT에게 보낼 텍스트 파싱 필요
 def word_preprocess(url : str):
@@ -30,8 +35,8 @@ def word_preprocess(url : str):
         
         # Step 1-1: image resizing 
         upload_image = upload_to_s3(og_title, og_image, image_content_type) # before resizing
-        time.sleep(5) # need to refactor
-        resized_image = download_from_s3(upload_image)
+        # time.sleep(5) # need to refactor
+        # resized_image = download_from_s3(upload_image)
         
         # Step 2: 단어 파싱하기
         find_all_p = soup.find_all('p')
@@ -66,5 +71,5 @@ def word_preprocess(url : str):
         
         f.close()
         
-        return [og_title, resized_image, keyword]
+        return [og_title, DEFAULT_IMAGE_URL, keyword]
