--- backend/lambda-s3/requirements.txt
@@ -1,1 +1,0 @@
-Pillow == 9.5.0
