--- backend/lambda-s3/template.yaml
@@ -1,19 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: AWS::Serverless-2016-10-31
-Resources:
-  CreateThumbnail:
-    Type: AWS::Serverless::Function
-    Properties:
-      Handler: lambda_function.lambda_handler
-      Runtime: python3.10
-      Timeout: 10
-      Policies: AWSLambdaExecute
-      Events:
-        CreateThumbnailEvent:
-          Type: S3
-          Properties:
-            Bucket: !Ref SrcBucket
-            Events: s3:ObjectCreated:*
-
-  SrcBucket:
-    Type: AWS::S3::Bucket
